adicionei um novo provedor de autênticação com google, o nextauth, dessa forma a chamada fica no lado do servidor impedindo que haja vazamentos para o lado do cliente em relação ao seu token de acesso que é usado para realizar a autenticação com google.
