+++ .env.example
@@ -1,6 +1,8 @@
-NEXT_PUBLIC_PUBLIC_KEY="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+GOOGLE_CLIENT_ID="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+GOOGLE_CLIENT_SECRET="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+NEXTAUTH_SECRET="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+NEXTAUTH_URL="http://localhost:3000"
+
+NEXT_PUBLIC_EMAILJS_PUBLIC_KEY="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
 NEXT_PUBLIC_TEMPLATE_ID="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
 NEXT_PUBLIC_SERVICE_ID="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
-NEXT_PUBLIC_SUPABASE_URL="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
-NEXT_PUBLIC_SUPABASE_ANON_KEY="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
-NEXT_PUBLIC_URL_REDIRECT="xxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
