+++ package.json
@@ -16,16 +16,17 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@headlessui/react": "^1.7.18",
     "@heroicons/react": "^2.1.1",
-    "dotenv": "^10.0.0",
-    "local-storage": "^2.0.0",
-    "next": "14.1.0",
-    "react": "^18",
-    "react-dom": "^18",
-    "sonner": "^1.4.0",
     "@react-oauth/google": "^0.12.1",
     "@supabase/auth-helpers-nextjs": "^0.9.0",
     "@supabase/auth-helpers-react": "^0.4.2",
-    "@supabase/supabase-js": "^2.39.7"
+    "@supabase/supabase-js": "^2.39.7",
+    "dotenv": "^10.0.0",
+    "local-storage": "^2.0.0",
+    "next": "14.1.0",
+    "next-auth": "^4.24.6",
+    "react": "^18",
+    "react-dom": "^18",
+    "sonner": "^1.4.0"
   },
   "devDependencies": {
     "@types/node": "^20",
