+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,6 @@
+import { authConfig } from "@/lib/auth";
+import NextAuth from "next-auth/next";
+
+const handler = NextAuth(authConfig);
+
+export { handler as GET, handler as POST };
