+++ types/next-auth.d.ts
@@ -1,0 +1,10 @@
+import NextAuth from 'next-auth'
+
+declare module 'next-auth' {
+  /**
+   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
+   */
+  interface Session {
+    accessToken?: string
+  }
+}
